requirements_mac.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.8.3
aiosignal==1.4.0
aniso8601==10.0.1
annotated-types==0.7.0
anyio==3.7.1
attrs==25.3.0
blinker==1.9.0
boto3==1.36.0
botocore==1.36.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.3.0
cmake==4.1.0
distro==1.9.0
filelock==3.19.1
Flask==2.3.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-RESTful==0.3.10
frozenlist==1.7.0
fsspec==2025.9.0
googlemaps==4.4.2
grpcio==1.75.0
grpcio-tools==1.75.0
h11==0.16.0
h2==4.3.0
hf-xet==1.1.10
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.27.2
huggingface-hub==0.16.4
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.11.0
jmespath==1.0.1
joblib==1.5.2
lit==18.1.8
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.6.4
networkx==3.5
nltk==3.9.1
numpy==1.26.4
oauth2client==4.1.3
openai==1.3.5
packaging==25.0
pillow==11.3.0
portalocker==2.10.1
propcache==0.3.2
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.15.1
RapidFuzz==3.12.2
regex==2024.11.6
requests==2.31.0
rsa==4.9
s3transfer==0.11.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==2.2.2
sentencepiece==0.2.1
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.33.2
twilio==9.4.3
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.20.1